@@ -1,30 +1,24 @@
-const { sequelize } = require('../models')
-const { tag: TagModel, article: ArticleModel, category: CategoryModel } = require('../models')
+const Joi = require('joi')
 
-module.exports = {
-  async getTags(ctx) {
+// import models
+const { tag: TagModel, category: CategoryModel, sequelize } = require('../models')
+
+class TagController {
+  static async getTagList(ctx) {
     const data = await TagModel.findAll({
       attributes: ['name', [sequelize.fn('COUNT', sequelize.col('name')), 'count']],
       group: 'name'
     })
-    ctx.body = { code: 200, data }
-  },
+    ctx.client(200, 'success', data)
+  }
 
-  async getArticlesByTag(ctx) {
-    let { page = 1, pageSize = 15, name } = ctx.query,
-      offset = (page - 1) * pageSize
-
-    pageSize = parseInt(pageSize)
-
-    const data = await ArticleModel.findAndCountAll({
-      attributes: ['id', 'title', 'createdAt'],
-      include: [{ model: TagModel, where: { name } }, { model: CategoryModel }],
-      offset,
-      limit: pageSize,
-      order: [['createdAt', 'DESC']],
-      distinct: true
+  static async getCategoryList(ctx) {
+    const data = await CategoryModel.findAll({
+      attributes: ['name', [sequelize.fn('COUNT', sequelize.col('name')), 'count']],
+      group: 'name'
     })
-
-    ctx.body = { code: 200, ...data }
+    ctx.client(200, 'success', data)
   }
 }
+
+module.exports = TagController
