@@ -1,164 +1,263 @@
 const Joi = require('joi')
-const UserSchema = require('../schemas/user')
-
+const axios = require('axios')
+const { GITHUB, ADMIN_GITHUB_LOGIN_NAME } = require('../config')
+const { decodeQuery } = require('../utils')
+const { comparePassword, encrypt } = require('../utils/bcrypt')
+const { createToken } = require('../utils/token')
 const { user: UserModel, comment: CommentModel, reply: ReplyModel, sequelize } = require('../models')
-const { encrypt, comparePassword } = require('../lib/bcrypt')
-const { createToken, checkAuth } = require('../lib/token')
 
-// const emailRegexp = /^[A-Za-z0-9\u4e00-\u9fa5]+@[a-zA-Z0-9_-]+(\.[a-zA-Z0-9_-]+)+$/
+/**
+ * 读取 github 用户信息
+ * @param {String} username - github 登录名
+ */
+async function getGithubInfo(username) {
+  const result = await axios.get(`${GITHUB.fetch_user}${username}`)
+  return result.data
+}
 
-module.exports = {
-  async register(ctx) {
-    const { username, password, email } = ctx.request.body
-    let response
-    const validator = Joi.validate({ username, password, email }, UserSchema.regiester)
-    if (validator.error) {
-      response = { code: 400, message: validator.error.message }
+class UserController {
+  // ===== utils methods
+  // 查找用户
+  static find(params) {
+    return UserModel.findOne({ where: params })
+  }
+
+  // 创建用户
+  static createGithubUser(data, role = 2) {
+    const { id, login, email } = data
+    return UserModel.create({
+      id,
+      username: login,
+      role,
+      email,
+      github: JSON.stringify(data)
+    })
+  }
+
+  // 更新用户信息
+  static updateUserById(userId, data) {
+    return UserModel.update(data, { where: { id: userId } })
+  }
+  // ===== utils methods
+
+  // 登录
+  static login = async ctx => {
+    const { code } = ctx.request.body
+    if (code) {
+      await this.githubLogin(ctx, code)
     } else {
+      await this.defaultLogin(ctx)
+    }
+  }
+
+  // 站内用户登录
+  static defaultLogin = async ctx => {
+    const validator = ctx.validate(ctx.request.body, {
+      account: Joi.string().required(),
+      password: Joi.string()
+    })
+    if (validator) {
+      try {
+        const { account, password } = ctx.request.body
+
+        const user = await UserModel.findOne({
+          where: {
+            // $or: { email: account, username: account }
+            username: account
+          }
+        })
+
+        if (!user) {
+          ctx.client(403, '用户不存在')
+        } else {
+          const isMatch = await comparePassword(password, user.password)
+          if (!isMatch) {
+            ctx.client(403, '密码不正确')
+          } else {
+            const { id, role } = user
+            const token = createToken({ username: user.username, userId: id, role }) // 生成 token
+            ctx.client(200, '登录成功', { username: user.username, role, userId: id, token })
+          }
+        }
+      } catch (error) {
+        console.log(error)
+        // ctx.client(500, null, error)
+        throw error
+      }
+    }
+  }
+
+  // github 登录
+  static githubLogin = async (ctx, code) => {
+    try {
+      const result = await axios.post(GITHUB.access_token_url, {
+        client_id: GITHUB.client_id,
+        client_secret: GITHUB.client_secret,
+        code
+      })
+
+      const { access_token } = decodeQuery(result.data)
+
+      if (access_token) {
+        // 拿到 access_token 去获取用户信息
+        const result2 = await axios.get(`${GITHUB.fetch_user_url}?access_token=${access_token}`)
+        const githubInfo = result2.data
+
+        let target = await this.find({ id: githubInfo.id }) // 在数据库中查找该用户是否存在
+
+        if (!target) {
+          target = await UserModel.create({
+            id: githubInfo.id,
+            username: githubInfo.name || githubInfo.username,
+            github: JSON.stringify(githubInfo),
+            email: githubInfo.email
+          })
+        } else {
+          if (target.github !== JSON.stringify(githubInfo)) {
+            // github 信息发生了变动
+            // console.log(`${githubInfo.login}: github 信息发生改变， 更新 user....`)
+            const { id, login, email } = githubInfo
+            const data = {
+              username: login,
+              email,
+              github: JSON.stringify(githubInfo)
+            }
+            await this.updateUserById(id, data)
+          }
+        }
+        // username: user.username, role, userId: id, token
+        const token = createToken({ userId: githubInfo.id, role: target.role }) // 生成 token
+
+        ctx.client(200, 'success', {
+          github: githubInfo,
+          username: target.username,
+          userId: target.id,
+          role: target.role,
+          token
+        })
+      } else {
+        ctx.client(403, 'github 授权码已失效！')
+      }
+    } catch (error) {
+      ctx.client(500, 'github 登录超时')
+    }
+  }
+
+  // 注册
+  static async register(ctx) {
+    const validator = ctx.validate(ctx.request.body, {
+      username: Joi.string().required(),
+      password: Joi.string().required(),
+      email: Joi.string()
+        .email()
+        .required()
+    })
+
+    if (validator) {
+      const { username, password, email } = ctx.request.body
       const result = await UserModel.findOne({ where: { email } })
       if (result) {
-        response = { code: 400, message: '邮箱已被注册' }
+        ctx.client(403, '邮箱已被注册')
       } else {
         const user = await UserModel.findOne({ where: { username } })
-        if (user) {
-          response = { code: 400, message: '用户名已被占用' }
+        if (user && !user.github) {
+          ctx.client(403, '用户名已被占用')
         } else {
           const saltPassword = await encrypt(password)
           await UserModel.create({ username, password: saltPassword, email })
-          response = { code: 200, message: '注册成功' }
+          ctx.client(200, '注册成功')
         }
       }
-    }
-    ctx.body = response
-  },
-
-  async login(ctx) {
-    const { account, password } = ctx.request.body
-    const validator = Joi.validate({ account, password }, UserSchema.login)
-    let response
-    if (validator.error) {
-      response = { code: 400, message: validator.error.message }
-    } else {
-      const user = await UserModel.findOne({
-        where: {
-          $or: { username: account, email: account }
-        }
-      })
-      if (!user) {
-        response = { code: 400, message: '用户不存在' }
-      } else {
-        const isMatch = await comparePassword(password, user.password)
-        if (!isMatch) {
-          response = { code: 400, message: '密码不正确' }
-        } else {
-          const { id, auth } = user
-          const token = createToken({ username: user.username, userId: id, auth, email: user.email }) // 生成 token
-          response = { code: 200, message: '登录成功', username: user.username, auth: user.auth, token }
-        }
-      }
-    }
-
-    ctx.body = response
-  },
-
-  /**
-   * 更新账户信息
-   *
-   */
-  async updateUser(ctx) {
-    const userId = parseInt(ctx.params.id) // userId
-    const { username, oldPassword, password, email } = ctx.request.body
-    const user = await UserModel.findByPk(userId) // 为什么不直接用 update ==> 防止不法分子直接把博主的密码改了，或者其他...
-
-    let response = {},
-      token
-
-    if (!user.email) {
-      const result = await UserModel.findOne({ where: { email } })
-      if (result) {
-        response = { code: 400, message: '该邮箱已被注册' }
-      } else {
-        await UserModel.update({ email }, { where: { id: userId } })
-        response = { code: 200, message: `已成功绑定邮箱 ${email}` }
-      }
-    } else {
-      if (oldPassword) {
-        const isMatch = await comparePassword(oldPassword, user.password)
-        if (isMatch) {
-          if (!username && !password) {
-            response = { code: 400, message: '用户名/密码参数错误' }
-          } else if (username && !password) {
-            const result = await UserModel.findOne({ where: { username } })
-            if (result) {
-              response = { code: 400, message: '用户名已被占用' }
-            } else {
-              await UserModel.update({ username }, { where: { id: userId } })
-              response = { code: 200, message: '用户名修改成功' }
-            }
-          } else if (!username && password) {
-            const saltPassword = await encrypt(password)
-            await UserModel.update({ password: saltPassword }, { where: { id: userId } })
-            response = { code: 200, message: '密码修改成功' }
-          } else if (username && password) {
-            const result = await UserModel.findOne({ where: { username } })
-            if (result && result.username !== username) {
-              response = { code: 400, message: '用户名已被占用' }
-            } else {
-              const saltPassword = await encrypt(password)
-              await UserModel.update({ username, password: saltPassword }, { where: { id: userId } })
-              response = { code: 200, message: '用户名/密码修改成功' }
-            }
-          }
-        } else {
-          response = { code: 400, message: '密码不正确' }
-        }
-      } else {
-        response = { code: 400, message: '请输入原密码验证您的身份' }
-      }
-    }
-
-    if (response.code === 200) {
-      const result = await UserModel.findById(userId)
-      const { username, id, email, auth } = result
-      token = createToken({ username, userId: id, auth, email }) // 生成 token
-      response.token = token
-    }
-    ctx.body = response
-  },
-
-  async getUserList(ctx) {
-    const isAuth = checkAuth(ctx)
-    if (isAuth) {
-      let { page = 1, pageSize = 10, username } = ctx.query
-      const offset = (page - 1) * pageSize
-      pageSize = parseInt(pageSize)
-
-      const params = username ? { username: { $like: `%${username}%` } } : {}
-
-      const data = await UserModel.findAndCountAll({
-        attributes: ['id', 'username', 'createdAt'],
-        where: { auth: 2, ...params },
-        include: [{ model: CommentModel, attributes: ['id'], include: [{ model: ReplyModel, attributes: ['id'] }] }],
-        offset,
-        limit: pageSize,
-        row: true,
-        distinct: true,
-        order: [['createdAt', 'DESC']]
-      })
-      ctx.body = { code: 200, ...data }
-    }
-  },
-
-  async delete(ctx) {
-    const isAuth = checkAuth(ctx)
-    if (isAuth) {
-      let { userId } = ctx.query
-      userId = parseInt(userId)
-      await sequelize.query(
-        `delete comment, reply from comment left join reply on comment.id=reply.commentId where comment.userId=${userId}`
-      )
-      await UserModel.destroy({ where: { id: userId } })
-      ctx.body = { code: 200, message: '成功删除用户' }
     }
   }
+
+  /**
+   * 获取用户列表
+   */
+  static async getList(ctx) {
+    const validator = ctx.validate(ctx.query, {
+      username: Joi.string(),
+      page: Joi.string(),
+      pageSize: Joi.number()
+    })
+
+    if (validator) {
+      const { page = 1, pageSize = 10, username } = ctx.query
+      const where = {
+        username: {
+          $not: ADMIN_GITHUB_LOGIN_NAME
+        }
+      }
+      if (username) {
+        where.username['$like'] = `%${username}%`
+      }
+      const result = await UserModel.findAndCountAll({
+        where,
+        offset: (page - 1) * pageSize,
+        limit: parseInt(pageSize),
+        row: true
+      })
+
+      ctx.client(200, 'success', result)
+    }
+  }
+
+  static async delete(ctx) {
+    const validator = ctx.validate(ctx.params, {
+      userId: Joi.number().required()
+    })
+
+    if (validator) {
+      await sequelize.query(
+        `delete comment, reply from comment left join reply on comment.id=reply.commentId where comment.userId=${ctx.params.userId}`
+      )
+      await UserModel.destroy({ where: { id: ctx.params.userId } })
+      ctx.client(200)
+    }
+  }
+
+  /**
+   * 更新用户
+   */
+  static updateUser = async ctx => {
+    const validator = ctx.validate(
+      {
+        ...ctx.params,
+        ...ctx.request.body
+      },
+      {
+        userId: Joi.number().required(),
+        notice: Joi.boolean()
+      }
+    )
+
+    if (validator) {
+      const { userId } = ctx.params
+      const { notice } = ctx.request.body
+      await this.updateUserById(userId, { notice })
+      ctx.client(200)
+    }
+  }
+
+  /**
+   * 初始化用户
+   * @param {String} githubName - github name
+   */
+  static initGithubUser = async githubName => {
+    const temp = await this.find({ username: githubName })
+    if (!temp) {
+      const result = await getGithubInfo(githubName)
+      this.createGithubUser(result, 1)
+    }
+
+    // const tempList = await Promise.all(list.map(username => this.find({ username }))) // 查找库里是否有这个用户
+    // list.forEach(async (username, i) => {
+    //   if (!tempList[i]) {
+    //     const result = await getGithubInfo(username)
+    //     this.createGithubUser(result, 1)
+    //   }
+    // })
+  }
 }
+
+module.exports = UserController
