@@ -7,7 +7,7 @@ module.exports = (sequelize, dataTypes) => {
       id: { type: dataTypes.INTEGER(11), primaryKey: true, autoIncrement: true },
       title: { type: dataTypes.STRING(255), allowNull: false },
       content: { type: dataTypes.TEXT },
-      showOrder: { type: dataTypes.INTEGER(11) }, // 置顶文章展示的 order
+      viewCount: { type: dataTypes.INTEGER(11), defaultValue: 0 }, // 阅读数
       createdAt: {
         type: dataTypes.DATE,
         defaultValue: dataTypes.NOW,
