@@ -1,16 +1,53 @@
 const fs = require('fs')
 const path = require('path')
-const config = require('../config')
+const { DATABASE } = require('../config')
 const Sequelize = require('sequelize')
 
-const sequelize = new Sequelize(
-  config.db.database,
-  config.db.user,
-  config.db.password,
-  config.db.options
-)
+const Op = Sequelize.Op
 
-let db = {}
+const sequelize = new Sequelize(DATABASE.database, DATABASE.user, DATABASE.password, {
+  ...DATABASE.options,
+  // 在 sequelize V4 版本以后新加了符号运算符来代替 Op.xxx
+  // https://sequelize.org/master/manual/querying.html#operators
+  operatorsAliases: {
+    $eq: Op.eq,
+    $ne: Op.ne,
+    $gte: Op.gte,
+    $gt: Op.gt,
+    $lte: Op.lte,
+    $lt: Op.lt,
+    $not: Op.not,
+    $in: Op.in,
+    $notIn: Op.notIn,
+    $is: Op.is,
+    $like: Op.like,
+    $notLike: Op.notLike,
+    $iLike: Op.iLike,
+    $notILike: Op.notILike,
+    $regexp: Op.regexp,
+    $notRegexp: Op.notRegexp,
+    $iRegexp: Op.iRegexp,
+    $notIRegexp: Op.notIRegexp,
+    $between: Op.between,
+    $notBetween: Op.notBetween,
+    $overlap: Op.overlap,
+    $contains: Op.contains,
+    $contained: Op.contained,
+    $adjacent: Op.adjacent,
+    $strictLeft: Op.strictLeft,
+    $strictRight: Op.strictRight,
+    $noExtendRight: Op.noExtendRight,
+    $noExtendLeft: Op.noExtendLeft,
+    $and: Op.and,
+    $or: Op.or,
+    $any: Op.any,
+    $all: Op.all,
+    $values: Op.values,
+    $col: Op.col
+  }
+})
+
+const db = {}
 
 fs.readdirSync(__dirname)
   .filter(file => file !== 'index.js')
