@@ -1,47 +1,51 @@
 import React, { useEffect } from 'react'
 import { BrowserRouter, Route, Switch, Redirect } from 'react-router-dom'
-import routes from '@/routes/config'
 import { connect } from 'react-redux'
-import { getTags, getCategories } from '@/redux/article/actions'
-import { getWindowWidth } from '@/redux/common/actions'
 
-function Root(props) {
+import routes from '@/routes'
+import { getWindowWidth } from '@/redux/app/actions'
+import { getTagList, getCategoryList } from '@/redux/article/actions'
+
+const App = props => {
   // 初始化数据 类似 componentDidMount
   useEffect(() => {
-    props.getTags()
-    props.getCategories()
     props.getWindowWidth()
+    props.getTagList()
+    props.getCategoryList()
+    //
+    console.log('app did mount')
+    /*eslint react-hooks/exhaustive-deps: "off"*/
   }, [])
 
+  // 解构 route
   function renderRoutes(routes, contextPath) {
     const children = []
 
     const renderRoute = (item, routeContextPath) => {
       let newContextPath = item.path ? `${routeContextPath}/${item.path}` : routeContextPath
       newContextPath = newContextPath.replace(/\/+/g, '/')
-
-      // auth handler
-      if ((item.protected || newContextPath.includes('admin')) && props.auth !== 1) {
+      if (newContextPath.includes('admin') && props.role !== 1) {
         item = {
           ...item,
-          component: () => <Redirect to="/login" />,
+          component: () => <Redirect to='/' />,
           children: []
         }
       }
 
-      if (item.component && item.childRoutes) {
-        const childRoutes = renderRoutes(item.childRoutes, newContextPath)
-        children.push(
-          <Route
-            key={newContextPath}
-            render={props => <item.component {...props}>{childRoutes}</item.component>}
-            path={newContextPath}
-          />
-        )
-      } else if (item.component) {
-        children.push(<Route key={newContextPath} component={item.component} path={newContextPath} exact />)
-      } else if (item.childRoutes) {
-        item.childRoutes.forEach(r => renderRoute(r, newContextPath))
+      if (item.component) {
+        if (item.childRoutes) {
+          const childRoutes = renderRoutes(item.childRoutes, newContextPath)
+          children.push(
+            <Route
+              key={newContextPath}
+              render={props => <item.component {...props}>{childRoutes}</item.component>}
+              path={newContextPath}
+            />
+          )
+          item.childRoutes.forEach(r => renderRoute(r, newContextPath))
+        } else {
+          children.push(<Route key={newContextPath} component={item.component} path={newContextPath} exact />)
+        }
       }
     }
 
@@ -54,74 +58,40 @@ function Root(props) {
   return <BrowserRouter>{children}</BrowserRouter>
 }
 
-// connect 不支持 function component 的装饰器写法 so...
 export default connect(
-  state => state.user,
-  { getTags, getCategories, getWindowWidth }
-)(Root)
+  state => ({
+    role: state.user.role
+  }),
+  { getWindowWidth, getTagList, getCategoryList }
+)(App)
 
-// @connect(
-//   state => state.user,
-//   { getTags, getCategories, getWindowWidth }
-// )
-// class Root extends Component {
-//   static defaultProps = {
-//     isLogin: false
-//   }
+// example test
+// import WebLayout from '@/layout/web'
+// import AdminLayout from '@/layout/admin'
+// import Example from '@/views/examples'
 
-//   componentDidMount() {
-//     this.props.getTags()
-//     this.props.getCategories()
-//     this.props.getWindowWidth()
-//   }
-
-
-//   /**
-//    * 根据路由表生成路由组件
-//    * @param {Array} routes - 路由配置表
-//    * @param {String} contextPath - 父级路径。比如后台 admin...
-//    */
-//   renderRoutes(routes, contextPath) {
-//     const children = []
-
-//     const renderRoute = (item, routeContextPath) => {
-//       let newContextPath = item.path ? `${routeContextPath}/${item.path}` : routeContextPath
-//       newContextPath = newContextPath.replace(/\/+/g, '/')
-
-//       // auth handler
-//       if ((item.protected || newContextPath.includes('admin')) && this.props.auth !== 1) {
-//         item = {
-//           ...item,
-//           component: () => <Redirect to="/login" />,
-//           children: []
-//         }
-//       }
-
-//       if (item.component && item.childRoutes) {
-//         const childRoutes = this.renderRoutes(item.childRoutes, newContextPath)
-//         children.push(
-//           <Route
-//             key={newContextPath}
-//             render={props => <item.component {...props}>{childRoutes}</item.component>}
-//             path={newContextPath}
-//           />
-//         )
-//       } else if (item.component) {
-//         children.push(<Route key={newContextPath} component={item.component} path={newContextPath} exact />)
-//       } else if (item.childRoutes) {
-//         item.childRoutes.forEach(r => renderRoute(r, newContextPath))
-//       }
-//     }
-
-//     routes.forEach(item => renderRoute(item, contextPath))
-
-//     return <Switch>{children}</Switch>
-//   }
-
-//   render() {
-//     const children = this.renderRoutes(routes, '/')
-//     return <BrowserRouter>{children}</BrowserRouter>
-//   }
+// const App = () => {
+//   return (
+//     <BrowserRouter>
+//       <Switch>
+//         <Route path='/test' component={Example} />
+//         <Route
+//           path='/admin'
+//           render={props => (
+//             <AdminLayout>
+//               <Route path='/admin/e' component={Example} exact />
+//             </AdminLayout>
+//           )}
+//         />
+//         <Route
+//           path='/'
+//           render={props => (
+//             <WebLayout>
+//               <Route path='/e' component={Example} />
+//             </WebLayout>
+//           )}
+//         />
+//       </Switch>
+//     </BrowserRouter>
+//   )
 // }
-
-// export default Root
