@@ -1,16 +1,23 @@
 import React from 'react'
-import './index.less'
-import greenMan from '@/assets/404.png'
-import { Link } from 'react-router-dom'
-import { Icon } from 'antd'
 
-const PageNotFound = () => {
+import { Result, Button } from 'antd'
+
+function PageNotFound(props) {
   return (
-    <div className="nofound-wrapper">
-      <img src={greenMan} alt="" className="green-man" />
-      <i className="iconfont icon-back" style={{ color: 'green', marginRight: 6 }} />
-      <Link to="/">返回首页</Link>
-    </div>
+    <Result
+      status='404'
+      title='404'
+      subTitle='Sorry, the page you visited does not exist.'
+      extra={
+        <Button
+          type='primary'
+          onClick={() => {
+            props.history.push('/')
+          }}>
+          Back Home
+        </Button>
+      }
+    />
   )
 }
 
