@@ -1,25 +1,37 @@
-import * as constants from '@/redux/constants'
+import * as TYPES from '@/redux/types'
+import { COLOR_LIST } from '@/utils/config'
+import { randomIndex } from '@/utils'
+
+// 生成 color
+function genertorColor(list = [], colorList = []) {
+  const _list = [...list].sort((x, y) => y.count - x.count)
+  _list.forEach((l, i) => {
+    l.color = colorList[i] || colorList[randomIndex(colorList)]
+  })
+  return _list
+}
 
 // state
 const defaultState = {
   categoryList: [],
-  tagList: [],
-  recentList: []
+  tagList: []
 }
 
 // reducer
-export const demoReducer = (state = defaultState, action) => {
+export const articleReducer = (state = defaultState, action) => {
   const { type, payload } = action
   switch (type) {
-    case constants.CATEGORY_GETLIST:
-      return { ...state, categoryList: payload }
+    case TYPES.ARTICLE_GET_TAG_LIST:
+      const tagList = genertorColor(payload, COLOR_LIST)
+      return { ...state, tagList }
 
-    case constants.TAG_GETLIST:
-      return { ...state, tagList: payload }
+    case TYPES.ARTICLE_GET_CATEGORY_LIST:
+      const categoryList = genertorColor(payload, COLOR_LIST)
+      return { ...state, categoryList }
 
     default:
       return state
   }
 }
 
-export default demoReducer
+export default articleReducer
