@@ -2,34 +2,28 @@ import React, { Component } from 'react'
 import './index.less'
 import { connect } from 'react-redux'
 import { Badge, Tag } from 'antd'
-
 import { Link } from 'react-router-dom'
 
-@connect(state => ({
-  categoryList: state.article.categoryList,
-  colorList: state.common.colorList
-}))
-class Categories extends Component {
-  render() {
-    const { categoryList, colorList } = this.props
+function Categories(props) {
+  const { categoryList } = props
+  return (
+    <div className='app-categories'>
+      <h2 className='title'>Categories</h2>
+      <p className='category-all-title'>{`${categoryList.length} categories in total`}</p>
 
-    return (
-      <div className="content-inner-wrapper categories">
-        <h2 className="title">Categories</h2>
-        <p className="category-all-title">{`${categoryList.length} categories in total`}</p>
-
-        <div className="categories-list">
-          {categoryList.map((item, i) => (
-            <Badge count={item.count} key={item.name}>
-              <Tag color={colorList[i]}>
-                <Link to={`/categories/${item.name}`}>{item.name}</Link>
-              </Tag>
-            </Badge>
-          ))}
-        </div>
+      <div className='categories-list'>
+        {categoryList.map((item, i) => (
+          <Badge count={item.count} key={item.name}>
+            <Tag color={item.color}>
+              <Link to={`/categories/${item.name}`}>{item.name}</Link>
+            </Tag>
+          </Badge>
+        ))}
       </div>
-    )
-  }
+    </div>
+  )
 }
 
-export default Categories
+export default connect(state => ({
+  categoryList: state.article.categoryList
+}))(Categories)
